# docs/swagger.yaml
definitions:
...
  pb.ApiResponseMonthSaldoBalances:
    properties:
      data:
        items:
          $ref: '#/definitions/pb.SaldoMonthBalanceResponse'
        type: array
      message:
        type: string
      status:
        type: string
    type: object
  pb.ApiResponseMonthTotalSaldo:
    properties:
      data:
        items:
          $ref: '#/definitions/pb.SaldoMonthTotalBalanceResponse'
        type: array
      message:
        type: string
      status:
        type: string
    type: object
...
  pb.ApiResponsePaginationSaldo:
    properties:
      data:
        items:
          $ref: '#/definitions/pb.SaldoResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/pb.PaginationMeta'
      status:
        type: string
    type: object
...
  pb.ApiResponseSaldo:
    properties:
      data:
        $ref: '#/definitions/pb.SaldoResponse'
      message:
        type: string
      status:
        type: string
    type: object
...
  pb.ApiResponseYearSaldoBalances:
    properties:
      data:
        items:
          $ref: '#/definitions/pb.SaldoYearBalanceResponse'
        type: array
      message:
        type: string
      status:
        type: string
    type: object
...
  pb.ApiResponsesSaldo:
    properties:
      data:
        items:
          $ref: '#/definitions/pb.SaldoResponse'
        type: array
      message:
        type: string
      status:
        type: string
    type: object
...
  requests.CreateSaldoRequest:
    properties:
      card_number:
        type: string
      total_balance:
        type: integer
    required:
    - card_number
    - total_balance
    type: object
...
  response.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/saldo/yearly-balances:
    get:
      consumes:
      - application/json
      description: Retrieve yearly saldo balances for a specific year.
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Yearly saldo balances
          schema:
            $ref: '#/definitions/pb.ApiResponseYearSaldoBalances'
        "400":
          description: Invalid year parameter
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve yearly saldo balances
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get yearly saldo balances
      tags:
      - Saldo
  /api/saldos:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all saldo data with pagination and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of saldo data
          schema:
            $ref: '#/definitions/pb.ApiResponsePaginationSaldo'
        "500":
          description: Failed to retrieve saldo data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all saldo data
      tags:
      - Saldo
  /api/saldos/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a saldo by its ID
      parameters:
      - description: Saldo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Saldo data
          schema:
            $ref: '#/definitions/pb.ApiResponseSaldo'
        "400":
          description: Invalid saldo ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve saldo data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find a saldo by ID
      tags:
      - Saldo
  /api/saldos/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all active saldo data
      produces:
      - application/json
      responses:
        "200":
          description: List of saldo data
          schema:
            $ref: '#/definitions/pb.ApiResponsesSaldo'
        "500":
          description: Failed to retrieve saldo data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve all active saldo data
      tags:
      - Saldo
  /api/saldos/card_number/{card_number}:
    get:
      consumes:
      - application/json
      description: Retrieve a saldo by its card number
      parameters:
      - description: Card number
        in: path
        name: card_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Saldo data
          schema:
            $ref: '#/definitions/pb.ApiResponseSaldo'
        "500":
          description: Failed to retrieve saldo data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find a saldo by card number
      tags:
      - Saldo
  /api/saldos/create:
    post:
      consumes:
      - application/json
      description: Create a new saldo record with the provided card number and total
        balance.
      parameters:
      - description: Create Saldo Request
        in: body
        name: CreateSaldoRequest
        required: true
        schema:
          $ref: '#/definitions/requests.CreateSaldoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created saldo record
          schema:
            $ref: '#/definitions/pb.ApiResponseSaldo'
        "400":
          description: 'Bad Request: Invalid request body or validation error'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create saldo
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new saldo
      tags:
      - Saldo
  /api/saldos/monthly-balances:
    get:
      consumes:
      - application/json
      description: Retrieve monthly saldo balances for a specific year.
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Monthly saldo balances
          schema:
            $ref: '#/definitions/pb.ApiResponseMonthSaldoBalances'
        "400":
          description: Invalid year parameter
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve monthly saldo balances
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get monthly saldo balances
      tags:
      - Saldo
  /api/saldos/monthly-total-balance:
    get:
      consumes:
      - application/json
      description: Retrieve the total saldo balance for a specific month and year.
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Monthly total saldo balance
          schema:
            $ref: '#/definitions/pb.ApiResponseMonthTotalSaldo'
        "400":
          description: Invalid year or month parameter
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve monthly total saldo balance
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get monthly total saldo balance
      tags:
      - Saldo
  /api/saldos/permanent/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete an existing saldo record by its ID.
      parameters:
      - description: Saldo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted saldo record
          schema:
            $ref: '#/definitions/pb.ApiResponseSaldoDelete'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to delete saldo
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete a saldo
      tags:
      - Saldo
  /api/saldos/permanent/all:
    post:
      consumes:
      - application/json
      description: Permanently delete all saldo records from the database.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted all saldo records permanently
          schema:
            $ref: '#/definitions/pb.ApiResponseSaldoAll'
        "500":
          description: Failed to permanently delete all saldo records
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete all saldo records
      tags:
      - Saldo
  /api/saldos/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore an existing saldo record from the trash by its ID.
      parameters:
      - description: Saldo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored saldo record
          schema:
            $ref: '#/definitions/pb.ApiResponseSaldo'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore saldo
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed saldo
      tags:
      - Saldo
  /api/saldos/restore/all:
    post:
      consumes:
      - application/json
      description: Restore all saldo records that were previously deleted.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored all saldo records
          schema:
            $ref: '#/definitions/pb.ApiResponseSaldoAll'
        "500":
          description: Failed to restore all saldo records
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore all saldo records
      tags:
      - Saldo
  /api/saldos/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all trashed saldo data
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed saldo data
          schema:
            $ref: '#/definitions/pb.ApiResponsesSaldo'
        "500":
          description: Failed to retrieve saldo data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed saldo data
      tags:
      - Saldo
  /api/saldos/trashed/{id}:
    post:
      consumes:
      - application/json
      description: Soft delete an existing saldo record by its ID.
      parameters:
      - description: Saldo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully trashed saldo record
          schema:
            $ref: '#/definitions/pb.ApiResponseSaldo'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to trashed saldo
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Soft delete a saldo
      tags:
      - Saldo
...
  /api/saldos/yearly-total-balance:
    get:
      consumes:
      - application/json
      description: Retrieve the total saldo balance for a specific year.
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Yearly total saldo balance
          schema:
            $ref: '#/definitions/pb.ApiResponseYearTotalSaldo'
        "400":
          description: Invalid year parameter
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve yearly total saldo balance
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get yearly total saldo balance
      tags:
      - Saldo
...
swagger: "2.0"